Refactored view and added command for sitemap generation
