+++ README.md
@@ -1,9 +1,12 @@
 [![Tests](https://github.com//ckanext-sitemap/workflows/Tests/badge.svg?branch=main)](https://github.com//ckanext-sitemap/actions)
 
 # ckanext-sitemap
+
 A CKAN extension that generates a sitemap XML file is designed to create a structured map of a CKAN instance's datasets and resources, making it easier for search engines to discover and index the available data. !
 
-## Installation
+## Getting Started
+
+### Installation
 
 **TODO:** Add any additional install steps to the list below.
    For example installing any non-Python dependencies or adding any required
@@ -17,10 +20,12 @@ To install ckanext-sitemap:
 
 2. Clone the source and install it on the virtualenv
 
+    ```bash
     git clone https://github.com//ckanext-sitemap.git
     cd ckanext-sitemap
     pip install -e .
     pip install -r requirements.txt
+    ```
 
 3. Add `sitemap` to the `ckan.plugins` setting in your CKAN
    config file (by default the config file is located at
@@ -28,9 +33,11 @@ To install ckanext-sitemap:
 
 4. Restart CKAN. For example if you've deployed CKAN with Apache on Ubuntu:
 
-     sudo service apache2 reload
+    ```bash
+    sudo service apache2 reload
+    ```
 
-## Configuration
+### Configuration
 
 You can configure this extension in the `production.ini` file of your CKAN instance. Ensure to set these environment variables according to your requirements for sitemap generation and management.
 
@@ -38,9 +45,36 @@ Environment Variable | Default Value | Description
 -------------------- | ------------- | -----------
 `ckanext.sitemap.directory` | [`./ckanext/sitemap/public`](./ckanext/sitemap/public/) | The directory path for storing generated sitemaps.
 `ckanext.sitemap.max_items` | `5000` | Maximum number of items per sitemap file. If the total count of resources exceeds this limit, the sitemap is split into multiple files.
+`ckanext.sitemap.autorenew` | `True` | If this option is enabled, the sitemaps will be automatically renewed whenever a user requests a sitemap and the existing sitemap is older than the Time-To-Live (TTL) value specified. Set this to False if you prefer a cron job to handle sitemap generation.
 `ckanext.sitemap.ttl` | `8 * 3600` (8 hours) | Time-To-Live (TTL) for sitemaps. Sitemaps older than this value (in seconds) are regenerated when a user visits a sitemap route.
-`ckanext.sitemap.include_resources` | `True` | Determines whether package resources (distributions) should be included in the sitemaps. Set to `True` to include resources, and `False` to exclude them.
-`ckanext.sitemap.include_language_alternatives` | `True` | Determines whether package resources (distributions) should be included in the sitemaps. Set to `True` to include resources, and `False` to exclude them.
+`ckanext.sitemap.resources` | `True` | Determines whether package resources (distributions) should be included in the sitemaps. Set to `True` to include resources, and `False` to exclude them.
+`ckanext.sitemap.language_alternatives` | `True` | Determines whether package resources (distributions) should be included in the sitemaps. Set to `True` to include resources, and `False` to exclude them.
+
+### Using Cron for Regular Sitemap Generation
+
+Using cron to generate sitemaps regularly can be advantageous, especially if the sitemap generation process is time-consuming.
+
+Ensure that the sitemap generation occurs within the time frame specified by `ckanext.sitemap.ttl`, or alternatively, set `ckanext.sitemap.autorenew` to `False` to prevent accidental triggering of sitemap generation by users.
+
+**Example Cron Job:**
+
+To schedule the command to run at 2 AM, 10 AM, and 6 PM:
+
+```bash
+0 2,10,18 * * * /usr/lib/ckan/default/bin/ckan -c /etc/ckan/default/production.ini ckanext-sitemap generate > /dev/null 2>&1
+```
+
+## Available Commands
+
+- `generate`
+
+    This command triggers the generation of the sitemap.
+
+    Usage:
+
+    ```plaintext
+    ckanext-sitemap generate
+    ```
 
 ## License
