+++ ckanext/sitemap/plugin.py
@@ -1,11 +1,13 @@
 import ckan.plugins as plugins
 import ckan.plugins.toolkit as toolkit
 import ckanext.sitemap.view as view
+from ckanext.sitemap import cli
 
 
 class SitemapPlugin(plugins.SingletonPlugin):
     plugins.implements(plugins.IConfigurer)
     plugins.implements(plugins.IBlueprint)
+    plugins.implements(plugins.IClick)
 
     # IConfigurer
     def update_config(self, config_):
@@ -16,6 +18,7 @@ class SitemapPlugin(plugins.SingletonPlugin):
     # IBlueprint
     def get_blueprint(self):
         return view.get_blueprints()
-        
 
-    
+    # IClick
+    def get_commands(self):
+        return cli.get_commands()
